# apps/api/api/openapi.yaml
openapi: 3.0.3
info:
  title: Pickle Go API
  description: |
    Pickle Go is a web-first pickleball event management platform.
    This API provides endpoints for user authentication, event management, and registration.
  version: 1.0.0
  contact:
    name: Pickle Go Team
    url: https://picklego.tw
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.picklego.tw/api/v1
    description: Production server
...
paths:
...
  /users/me:
    get:
      tags:
        - Users
      summary: Get Current User
      description: Get the authenticated user's profile
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/me/events:
    get:
      tags:
        - Users
      summary: Get My Events
      description: Get events hosted by the current user
      operationId: getMyEvents
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of hosted events
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          events:
                            type: array
                            items:
                              $ref: '#/components/schemas/Event'
                          total:
                            type: integer
  /users/me/registrations:
    get:
      tags:
        - Users
      summary: Get My Registrations
      description: Get events the current user has registered for
      operationId: getMyRegistrations
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of registrations
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          registrations:
                            type: array
                            items:
                              $ref: '#/components/schemas/RegistrationWithEvent'
                          total:
                            type: integer
...
  /events/{id}/register:
    post:
      tags:
        - Registrations
      summary: Register for Event
      description: Register the current user for an event
      operationId: registerEvent
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/RegistrationResponse'
        '400':
          description: Cannot register (already registered, event cancelled, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Registrations
      summary: Cancel Registration
      description: Cancel the current user's registration for an event
      operationId: cancelRegistration
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Registration cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '404':
          description: Registration not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /events/{id}/registrations:
    get:
      tags:
        - Registrations
      summary: Get Event Registrations
      description: Get all registrations for an event
      operationId: getEventRegistrations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of registrations
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          confirmed:
                            type: array
                            items:
                              $ref: '#/components/schemas/RegistrationWithUser'
                          waitlist:
                            type: array
                            items:
                              $ref: '#/components/schemas/RegistrationWithUser'
                          confirmed_count:
                            type: integer
                          waitlist_count:
                            type: integer
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
    SuccessMessage:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            message:
              type: string
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
        avatar_url:
          type: string
          format: uri
        email:
          type: string
          format: email
    SkillLevel:
      type: string
      enum:
        - beginner
        - intermediate
        - advanced
        - expert
        - any
    EventStatus:
      type: string
      enum:
        - open
        - full
        - cancelled
        - completed
    Location:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
        google_place_id:
          type: string
    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
        host:
          $ref: '#/components/schemas/User'
        title:
          type: string
        description:
          type: string
        event_date:
          type: string
          format: date
        start_time:
          type: string
          pattern: '^\d{2}:\d{2}$'
        end_time:
          type: string
          pattern: '^\d{2}:\d{2}$'
        location:
          $ref: '#/components/schemas/Location'
        capacity:
          type: integer
          minimum: 4
          maximum: 20
        confirmed_count:
          type: integer
        waitlist_count:
          type: integer
        skill_level:
          $ref: '#/components/schemas/SkillLevel'
        skill_level_label:
          type: string
        fee:
          type: integer
          minimum: 0
          maximum: 9999
        status:
          $ref: '#/components/schemas/EventStatus'
...
    RegistrationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - confirmed
            - waitlist
        waitlist_position:
          type: integer
        message:
          type: string
    RegistrationWithUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user:
          $ref: '#/components/schemas/User'
        registered_at:
          type: string
          format: date-time
        waitlist_position:
          type: integer
    RegistrationWithEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
        status:
          type: string
        waitlist_position:
          type: integer
        registered_at:
          type: string
          format: date-time
        event:
          type: object
          properties:
            id:
              type: string
              format: uuid
            title:
              type: string
            event_date:
              type: string
              format: date
            start_time:
              type: string
            location:
              type: string
            skill_level:
              type: string
            status:
              type: string